Moves EntryPoint to its own module
